Extend imports

+++ src/ligths/point.cpp
@@ -9,7 +9,7 @@ PointLight::PointLight(const PointLight& light)
 {
 }
 
-PointLight::PointLight(double _ls, const RGBColor& _c, const Point& _loc)
+PointLight::PointLight(float _ls, const RGBColor& _c, const Point& _loc)
 		: ls(_ls), c(_c), location(_loc)
 {
 }
